+++ packages/ui/index.ts
@@ -1,5 +1,6 @@
 import Bootstrap_PT from "./src/Bootstrap_PT"
 import '@/styles/main.scss'
+import type { ComponentResolver, ComponentResolveResult } from 'unplugin-vue-components/types';
 
 import ForgeAlert from "@/components/ForgeAlert.vue";
 import ForgeActionButton from "./src/components/ForgeActionButton.vue";
@@ -22,11 +23,14 @@ import ForgeScratchpad from "./src/components/ForgeScratchpad.vue";
 import ForgeStatWidget from "./src/components/ForgeStatWidget.vue";
 import ForgeStepper from "./src/components/ForgeStepper.vue";
 import ForgeTile from "./src/components/ForgeTile.vue";
-import ForgeTable from "./src/components/table/ForgeTable.vue"; 
+import ForgeTable from "./src/components/table/ForgeTable.vue";
+import ForgeTreeTable from "./src/components/ForgeTreeTable.vue"; 
 import ForgeFilterHeader from "./src/components/table/ForgeFilterHeader.vue";
 import ForgeFileUploader from "./src/components/file-uploader/ForgeFileUploader.vue";
 import ForgeColumnCustomiser from './src/components/table/ForgeColumnCustomiser.vue'
 import ForgePanel from './src/components/ForgePanel.vue'
+import ForgeSelectButton from './src/components/ForgeSelectButton.vue'
+import ForgeToggleButton from './src/components/ForgeToggleButton.vue'
 import ForgeConfirmDialog from './src/components/ForgeConfirmDialog.vue'
 
 export {
@@ -57,8 +61,26 @@ export {
   ForgeFilterHeader,
   ForgeFileUploader,
   ForgePanel,
+  ForgeSelectButton,
+  ForgeToggleButton,
+  ForgeTreeTable,
   ForgeConfirmDialog
 }
 
 export * from './src/types/forge-types.ts'
 export * from "./src/helpers/index";
+
+export const ForgeResolver = (): ComponentResolver => {
+  return {
+    type: 'component',
+    resolve: (name: string) => {
+      if (!name.startsWith('Forge')) return;
+
+      return {
+        from: `@3squared/forge-ui-3`,
+        name: name
+      } as ComponentResolveResult;
+    }
+  };
+}
+