+++ packages/ui/src/Bootstrap_PT.ts
@@ -30,7 +30,11 @@ import CardPT from "./passthroughs/Card.pt.ts";
 import DividerPT from "./passthroughs/Divider.pt.ts";
 import ImagePT from "./passthroughs/Image.pt.ts";
 import ToastPT from "./passthroughs/Toast.pt.ts";
+import TreePT from "./passthroughs/Tree.pt.ts";
+import TreeSelectPT from "./passthroughs/TreeSelect.pt.ts";
 import FieldsetPt from "./passthroughs/Fieldset.pt.ts";
+import AccordionPT from "./passthroughs/Accordion.pt.ts";
+import TreeTable from "./passthroughs/TreeTable.pt.ts";
 
 export default {
   ...AutoCompletePT,
@@ -64,6 +68,10 @@ export default {
   ...ToastPT,
   ...TextAreaPT,
   ...FieldsetPt,
+  ...TreeSelectPT,
+  ...TreePT,
+  ...AccordionPT,
+  ...TreeTable,
   directives: {
     ...TooltipPT
   }