+++ packages/ui/src/components/ForgePanel.vue
@@ -4,10 +4,13 @@
       <Icon v-if="collapsed" data-cy="message-icon"  icon="bi:chevron-down"  />
       <Icon v-if="!collapsed" data-cy="message-icon"  icon="bi:chevron-up"  />
     </template>
-    <slot />
+    <template v-for="(_, slot) of $slots as unknown as PanelSlots" #[slot]="scope">
+      <slot :name="slot" v-bind="scope" />
+    </template>
   </Panel>
 </template>
 
 <script setup lang="ts">
 import { Icon } from "@iconify/vue";
+import { PanelSlots } from "primevue";
 </script>