+++ packages/ui/src/components/file-uploader/ForgeFileUploader.vue
@@ -3,7 +3,7 @@
     <UploadButton v-bind="props" v-model="files" />
     <div v-if="showDragDropArea">
       <DragDropArea :max-file-input="props.maxFileInput" :max-file-size="props.maxFileSize" :accepted-file-types="props.acceptedFileTypes" v-model="files">
-      <span v-for="({ file }, index) in files">
+      <div v-for="({ file }, index) in files">
         <FileInfo
             :key="file.name"
             :class="index === files.length - 1 || files.length == 1 ? '' : 'border-bottom'"
@@ -13,9 +13,9 @@
             v-bind="props"
             @deleted="deleteFiles"
         />
-      </span>
+      </div>
       </DragDropArea>
-      <MaxFileSize :max-file-size="props.maxFileSize"/>
+      <FileInputInfo :max-file-size="props.maxFileSize" :max-file-input="maxFileInput" />
     </div>
   </div>
 </template>
@@ -24,11 +24,11 @@
 import UploadButton from "@/components/file-uploader/components/UploadButton.vue";
 import DragDropArea from "@/components/file-uploader/components/DragDropArea.vue";
 import FileInfo from "@/components/file-uploader/components/FileInfo.vue";
-import MaxFileSize from "@/components/file-uploader/components/MaxFileSize.vue";
-import { ForgeFileStatus, ForgeFileType } from "../../types/forge-types";
-import {TypedSchema} from "vee-validate";
+import { ForgeFileStatus, ForgeFileType } from "@/types/forge-types.ts";
+import { TypedSchema } from "vee-validate";
+import FileInputInfo from "@/components/file-uploader/components/FileInputInfo.vue";
 
-const files = defineModel<ForgeFileStatus[]>({ default: []})
+const files = defineModel<ForgeFileStatus[]>({ default: [] })
 
 export interface ForgeFileUploaderProps {
   acceptedFileTypes: ForgeFileType[],
@@ -48,7 +48,7 @@ const props = withDefaults(defineProps<ForgeFileUploaderProps>(), {
   showDragDropArea: true
 })
 
-const deleteFiles = (fileInfo : File) => {
+const deleteFiles = (fileInfo: File) => {
   files.value = files.value.filter((f) => f.file != fileInfo)
 }
 </script>