+++ packages/ui/src/passthroughs/Accordion.pt.ts
@@ -1,0 +1,19 @@
+export default {
+  accordionPanel: {
+    root: "card my-2"
+  },
+  accordionHeader:{
+    root: "border-0 bg-transparent fw-400 w-100 d-flex flex-row justify-content-between align-content-center ps-4 pe-3 py-2 my-2",
+    //header: "h4 d-flex justify-content-between align-items-center fw-600 ms-4 me-3 my-2",
+    pcToggleButton: {
+      root: {
+        class: [
+          "btn",
+        ]
+      }
+    },
+  },
+  accordionContent: {
+    root: "mt-2 mx-4 mb-4"
+  }
+}