+++ packages/ui/src/passthroughs/Column.pt.ts
@@ -8,25 +8,26 @@ export default {
         class: [
           'align-top',
           {
-          'cursor-move': options.props.reorderableColumns,
-          'position-sticky': options.column?.context.frozen !== undefined || options.column?.context.frozen || options.column?.context.frozen === '',
-          'overflow-hidden position-relative bg-clip-padding': options.column?.context.resizable
-        }],
+            'cursor-move': options.props.reorderableColumns,
+            'position-sticky': options.column?.context.frozen !== undefined || options.column?.context.frozen || options.column?.context.frozen === '',
+            'overflow-hidden position-relative bg-clip-padding': options.column?.context.resizable
+          }],
       }
     },
-    
     pcRowEditorInit: 'btn',
     rowToggleButton: 'btn',
+    nodeToggleButton: "btn btn-sm",
     columnResizer: {
       class: 'position-absolute top-0 end-0 m-0 h-100 p-0 cursor-resize border border-transparent'
     },
     pcRowEditorSave: 'btn',
     pcRowEditorCancel: 'btn',
-    bodyCell: (options: ColumnPassThroughMethodOptions & { props: { resizableColumns: boolean }, column: { props: { frozen: boolean | '' } } }) => {
+    bodyCell: (options: ColumnPassThroughMethodOptions & { props: { resizableColumns: boolean }, column: { props: { frozen: boolean | '', expander: boolean | undefined } } }) => {
       return {
         class: {
+          'expander-cell': options.column.props?.expander,
           'position-sticky': options.column.props?.frozen !== undefined,
-          'overflow-hidden text-nowrap': options.props.resizableColumns
+          'overflow-hidden text-nowrap': options.parent.props.resizableColumns
         }
       }
     },
@@ -39,11 +40,11 @@ export default {
           }],
       }
     },
-    pcColumnFilterClearButton: ({ props } : ColumnPassThroughMethodOptions & { props: { filters: object}}) => {
+    pcColumnFilterClearButton: ({ props }: ColumnPassThroughMethodOptions & { props: { filters: object } }) => {
       return {
         class: [
           'btn',
-          { 
+          {
             'd-none': props.filters[props.field] ? props.filters[props.field]?.value === null : false
           }
         ]
@@ -62,7 +63,7 @@ export default {
     filter: 'd-flex w-100',
     filterElementContainer: "w-100",
     filterConstraintList: 'ps-0',
-    filterConstraint: (options : ColumnPassThroughMethodOptions) => {
+    filterConstraint: (options: ColumnPassThroughMethodOptions) => {
       return {
         class: [
           'dropdown-item cursor-pointer',
@@ -102,6 +103,6 @@ export default {
         ]
       }
     },
-  
+
   }
 }